ADD & UPDATE : Menambahkan komentar dan juga mengubah pada deleteData() menambahkan validasi

+++ app/Controllers/Http/UsersController.ts
@@ -9,13 +9,16 @@ export default class UsersController {
     }
 
     public async addData({request, response}) {
-
-        await request.validate(SimpleBlogValidator)
-
+        // Instantiate Object from model user
         var user = new User()
 
+        // Validation Input
+        await request.validate(SimpleBlogValidator)
+        
+        // User insert data using object
         await user.fill(request.all()).save()
 
+        // true
         if(user.$isPersisted){
             return response.json({
                 'response' : 200,
@@ -23,6 +26,7 @@ export default class UsersController {
             })
         }
 
+        // false
         return response.json({
             'response' : 404,
             'message' : "Error While Insert Data"
@@ -34,12 +38,20 @@ export default class UsersController {
     // }
 
     public async deleteData({request, response}){
-        var user = await User.findOrFail(request.param('id'))
+        var user = await User.find(request.param('id'))
+        
+        if(!user) {
+            return response.json({
+                'response' : 404,
+                'message' : "Data Not Found"
+            })
+        }
+        
         await user.delete()
-
+        
         return response.json({
             'response' : 200,
-            'message' : "Successfully deleted"
+            'message' : user
         })
     }
 }
